♻️ Refactor: 폴더 구조 재정리
컴포넌트와 그 컴포넌트의 스타일 파일을 같은 폴더로 정리
