+++ src/App.jsx
@@ -1,27 +1,12 @@
+import "./App.styles";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { lazy, Suspense } from "react";
-import styled from "@emotion/styled";
-import Lottie from "lottie-react";
 import loadingAnimation from "./assets/animations/loading-animation.json";
+import { StyledLottie, Wrapper } from "./App.styles";
 
 const Layout = lazy(() => import("./layout/Layout"));
-const Home = lazy(() => import("./pages/Home"));
+const Home = lazy(() => import("./pages/Home/Home"));
 const MovieDetail = lazy(() => import("./components/MovieDetail/MovieDetail"));
-
-const Wrapper = styled.div`
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: #eee;
-  font-size: 3rem;
-  font-weight: 600;
-`;
-
-const StyledLottie = styled(Lottie)`
-  width: 10rem;
-  height: 10rem;
-`;
 
 function App() {
   return (
