+++ src/App.styles.jsx
@@ -1,0 +1,17 @@
+import styled from "@emotion/styled";
+import Lottie from "lottie-react";
+
+export const Wrapper = styled.div`
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  color: #eee;
+  font-size: 3rem;
+  font-weight: 600;
+`;
+
+export const StyledLottie = styled(Lottie)`
+  width: 10rem;
+  height: 10rem;
+`;
